+++ libs/modules-fe/src/landing/banner-wellcome/index.ts
@@ -1,0 +1,1 @@
+export * from './section';
