+++ libs/modules-fe/src/landing/benefit/type.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+
+export type TBenefit = {
+  icon: ReactNode;
+  title: string;
+  desc: ReactNode;
+};
