+++ libs/modules-fe/src/landing/hero/index.ts
@@ -1,0 +1,1 @@
+export * from './section';
