+++ libs/modules-fe/src/landing/index.ts
@@ -1,2 +1,2 @@
 export * from './module';
-export * from './type';
+export * from './program-study/type';
