+++ libs/modules-fe/src/landing/program-study/index.ts
@@ -1,0 +1,1 @@
+export * from './section';
