+++ libs/modules-fe/src/landing/program-study/type.ts
