+++ package.json
@@ -71,6 +71,7 @@
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
     "react": "18.2.0",
+    "react-alice-carousel": "^2.7.1",
     "react-data-table-component": "^7.5.3",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
